Updated parser and related code to support TON syntax
This commit updates the parser and rewrites the desugaring code to bring the
supported language in line with the language described in the Merlin TON journal
paper.

The Merlin syntax is now composed of three elements:

  1. Mappings from names to sets of names, useful for defining abstract
  functions and groups of network entities.
  2. Core language policies, composed of triples of (variable, predicate, regex)
  and Presburger formulas over a particular set of triples.
  3. Macro language blocks, composed of a for-each iterator over some collection
  of locations, and associated predicate, regex and rate specification. Each
  such block desugars into a set of core language policies, with autogenerated
  variables and the rate spec used to produce a Presburger formula.

A Merlin program may have any or all of the above. The supporting code first
expands all macro blocks into core language policies, and then replaces names in
regexes with alternations over the appropriate set of network locations. Types
and functions have been changed and renamed to support the new language.

This new language and parsing strategy allows us to support the example policies
in the TON paper, with some superficial syntactic differences (`and` and `or`
instead of & and |, `ipSrc` instead of `ip.src` etc.).

The parser check if locations used in policies are actually available in the
given topology. Some portions of Merlin_FrontEnd.ml should also be rewritten to
be cleaner.