Fix compilation of policies without minimum guarantees
Policies without a minimum rate guarantee do not require use of the solver and
can be compiled using only the cross product of the regular expression and the
topology graph. This commit fixes how policies are partition based on their
associated bandwidth formula, and correctly passes associated the rate
information to the back-end for proper queue configuration and tc generation.